+++ internal/backend/init/init_test.go
@@ -29,6 +29,7 @@ func TestInit_backend(t *testing.T) {
 		{"inmem", "*inmem.Backend"},
 		{"pg", "*pg.Backend"},
 		{"s3", "*s3.Backend"},
+		{"orcle-oci", "*oracle_oci.Backend"},
 	}
 
 	// Make sure we get the requested backend
