+++ internal/backend/remote-state/oracle_oci/util.go
@@ -1,7 +1,7 @@
 // Copyright (c) The OpenTofu Authors
 // SPDX-License-Identifier: MPL-2.0
 
-package oci
+package oracle_oci
 
 import (
 	"fmt"
